requirements.txt
streamlit>=1.31
google-genai>=0.6
pillow>=10.2
python-dotenv>=1.0

